# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.oas36ty.test:8000/v1",
            "description": "Localhost"
        }
    ],
    "paths": {
...
        "/switch": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Switch",
                "description": "Switch",
                "operationId": "postSwitch",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/tenant--header"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tenant_id": {
                                        "type": "string",
                                        "example": "XXXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Message!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error Message!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized Error Message!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "tenant--header": {
                "name": "X-Tenant",
                "in": "header",
                "description": "Tenant ID",
                "schema": {
                    "type": "string",
                    "default": "oas36ty"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Execute Login or Signup APIs to get the authentication token",
                "name": "Token Based Authentication",
                "scheme": "bearer"
            }
        }
    }
}